Reorganize code and show new todo/project on creation
