+++ src/js/display/components/project_display.js
