+++ src/js/display/display_results.js
@@ -1,10 +1,8 @@
-// eslint-disable-next-line import/extensions
-import Accordion from 'accordion/src/accordion.mjs';
-import { createAccordions, closeFold, openFold } from './accordion_utils';
-import { getAppData } from './data';
-import { deleteProject, deleteTodo } from './methods';
+import { createAccordions } from './accordion_utils';
+import { getAppData, deleteProject, deleteTodo } from '../models/data';
 import { addTodoBtnAction, editTodoBtnAction } from './button_action';
-import createProjectDiv from './project_display';
+import createProjectDiv from './components/project_display';
+import { deleteProjectFromUI, deleteTodoFromUI } from './change_dom';
 
 const createResultsSection = (projects) => {
   const listOfProjects = document.createElement('div');
@@ -15,29 +13,6 @@ const createResultsSection = (projects) => {
     listOfProjects.appendChild(projectDiv);
   });
   return listOfProjects;
-};
-
-const deleteProjectFromUI = (projectId) => {
-  const projectDiv = document.getElementById(`project-${projectId}`);
-  const projectFold = Accordion.getFold(projectDiv);
-  const { accordion } = projectFold;
-  const foldIndex = accordion.folds.findIndex(fold => fold === projectFold);
-  accordion.folds.splice(foldIndex, 1);
-  projectDiv.remove();
-  setTimeout(closeFold, 200, projectFold);
-};
-
-const deleteTodoFromUI = (todoId, projectId) => {
-  const todoDiv = document.getElementById(`project-${projectId}-todo-${todoId}`);
-  const todoFold = Accordion.getFold(todoDiv);
-  const { accordion } = todoFold;
-  const { parentFold } = accordion;
-  setTimeout(closeFold, 200, todoFold);
-  const foldIndex = accordion.folds.findIndex(fold => fold === todoFold);
-  accordion.folds.splice(foldIndex, 1);
-  todoDiv.remove();
-  setTimeout(closeFold, 400, parentFold);
-  setTimeout(openFold, 600, parentFold);
 };
 
 const setupBtnEventListeners = () => {
