+++ src/js/form_builder/components/notes_section.js
