+++ src/js/models/data.js
@@ -36,4 +36,30 @@ const getTodo = (todoId, projectId) => {
   return todoToUpdate;
 };
 
-export { getAppData, storeAppData, getTodo };
+const deleteProject = (projectId) => {
+  if (projectId !== 0) {
+    const { defaultProject, otherProjects } = getAppData();
+    const projectIndex = otherProjects.findIndex(project => project.getId() === projectId);
+    otherProjects.splice(projectIndex, 1);
+    storeAppData({ defaultProject, otherProjects });
+  }
+};
+
+const deleteTodo = (todoId, projectId) => {
+  const { defaultProject, otherProjects } = getAppData();
+  const project = projectId === 0
+    ? defaultProject
+    : otherProjects.find(project => project.getId() === projectId);
+  const todos = project.getTodos();
+  const todoIndex = todos.findIndex(todo => todo.getTodoId() === todoId);
+  todos.splice(todoIndex, 1);
+  storeAppData({ defaultProject, otherProjects });
+};
+
+export {
+  getAppData,
+  storeAppData,
+  getTodo,
+  deleteProject,
+  deleteTodo,
+};
